Size is not grandeur Task commit #2

+++ 0x00-hello_world/6-size.c
@@ -3,7 +3,6 @@
 /**
 * main - Entry point
 *
-* @
 * Return: Always 0 (Success)
 */
 int main(void)
@@ -14,11 +13,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a char: %d bytes(s)\n", sizeof(a));
+	printf("Size of an int: %d bytes(s)\n", sizeof(b));
+	printf("Size of a long int: %d bytes(s)\n", sizeof(c));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("Size of a float: %d byte(s)\n", sizeof(e));
 
 	return (0);
 }
